version update, moved NetChanges to their own files
