+++ project/source/Version.h
@@ -2,7 +2,7 @@
 
 //-----------------------------------------------------------------------------
 #define VERSION_MAJOR 0
-#define VERSION_MINOR 5
+#define VERSION_MINOR 6
 #define VERSION_PATCH 0
 
 //-----------------------------------------------------------------------------
