+++ project/source/game/SaveState.h
@@ -11,6 +11,9 @@ enum SAVE_VERSION
 	V_0_3 = 5, // 0.3
 	V_0_4 = 6, // 0.4/0.4.10/0.4.20
 	V_0_5 = 7, // 0.5
+	V_0_6 = 8, // 0.6
+
+	V_CURRENT = V_0_6
 };
 
 //-----------------------------------------------------------------------------
